fix(ccshuffle): unbiased range via Lemire mul+shr, CCBI int64_t UB, ccbi CMake registration
Three fixes:

1. ccshuffle.h — replace biased prng_next() \% i with unbiased
   _ccshuffle_range():
   • GCC/Clang/ICC: (__uint128_t)rnd * range >> 64 (mul+shr, no div)
   • MSVC x64:     _umul128 (same zero-div cost)
   • MSVC ARM64:   __umulh
   • Fallback:     rejection sampling (two div ops)
   Rejection prob ~range/2^64 — for any practical N, essentially zero.

2. ccbi.h — fix INT64_MIN round-trip UB in ccbi_to_int:
   -(int64_t)v on v=0x8000000000000000 triggers signed overflow UB
   → negation in uint64_t arithmetic instead.

3. CMakeLists.txt — register test_ccbi in ctest + check target deps.
  It was previously only buildable manually.

docs: add ccshuffle API ref, update index tools table

+++ include/ccbi.h
@@ -1759,7 +1759,10 @@ ccbi_to_int(const ccbi_t *z) CCBI_NOEXCEPT {
   if (zu == 0) return 0;
   uint64_t v = z->limbs[0];
   if (zu >= 2) v |= (uint64_t)z->limbs[1] << 32;
-  return (int64_t)(CCBI_SIGN(z) < 0 ? -(int64_t)v : (int64_t)v);
+  /* negation in unsigned arithmetic avoids UB on INT64_MIN;
+   * the final cast is implementation-defined on two's complement
+   * but yields the correct INT64_MIN on all real platforms. */
+  return (int64_t)(CCBI_SIGN(z) < 0 ? -v : v);
 }
 
 CCBI_INLINE int
